requirements.txt
python-ESL
sqlalchemy
Flask
Flask-Docs
Flask-SQLAlchemy
pymysql==0.10.0
paho-mqtt==1.5.0
celery==4.4.7
flask-cors==3.0.9